requirements.txt
fire==0.5.0
python-dotenv==1.0.0
flask==2.0.0
waitress==2.1.2
Flask-Limiter==3.4.0
werkzeug==2.3.7
redis==4.6.0
rich==13.7.0
cloudpickle==3.0.0